refactor(by_age_range): move to root
