--- cepimose/by_age_range/types.py
@@ -1,14 +1,0 @@
-from dataclasses import dataclass
-import datetime
-
-
-@dataclass
-class VaccinationDose:
-    date: datetime.datetime
-    dose: int
-
-
-@dataclass
-class VaccinationByAgeRange:
-    dose1: "list[VaccinationDose]"
-    dose2: "list[VaccinationDose]"
